tests: split mockers.py and rename mockers_git_repository.py
Most helpers in tests/mockers.py weren't actually mocks (CLI runner,
output assertions, shell/FS utilities, branch-layout-file readers), and
mockers_git_repository.py contained pure git-CLI wrappers - none of them
mocking anything. Split/rename so module names reflect contents:

- tests/mockers.py: trimmed to real stubs only (mock_*, env-override
  context managers fixed_author_and_committer_date_in_past,
  overridden_environment).
- tests/cli_runner.py (new): in-process git-machete runner + output
  assertions (launch_command, assert_success, assert_failure, ...) +
  branch-layout-file helpers.
- tests/shell.py (new): generic shell/FS utilities (execute, popen,
  read_file, write_to_file, ...).
- tests/mockers_git_repository.py -> tests/git_repository.py; also
  absorbed wait_to_bump_commit_timestamp from mockers.py.

Imports updated across all 30+ test files, plus tests/__init__.py
(register_assert_rewrite -> tests.cli_runner) and the
prohibit-sleep-in-python.sh check (allowlist now points at
tests/git_repository.py).

Also:
- Reorder teardown_method in BaseTest so the chdir into the empty temp
  dir runs *before* the patched-method-never-called assertion - that way
  a failing assertion can't leave a later test pinned to a stale cwd.
- Rewrite mock__run_cmd_and_forward_stdout's docstring: it patches
  git_machete.utils._run_cmd (not utils.run_cmd as previously claimed),
  it does NOT capture stderr (stderr=None means inherited), and it
  returns the exit code (the captured stdout reaches the test only
  indirectly via launch_command's redirect_stdout).
