+++ tests/completion_e2e/test_completion_e2e.py
@@ -4,10 +4,11 @@ from typing import Dict
 
 import pytest
 
-from tests.mockers import popen, rewrite_branch_layout_file
-from tests.mockers_git_repository import (check_out, commit, create_repo,
-                                          get_current_commit_hash, new_branch,
-                                          set_git_config_key)
+from tests.cli_runner import rewrite_branch_layout_file
+from tests.git_repository import (check_out, commit, create_repo,
+                                  get_current_commit_hash, new_branch,
+                                  set_git_config_key)
+from tests.shell import popen
 
 test_cases: Dict[str, str] = {
     "git machete ":
