+++ tests/test_add.py
@@ -2,13 +2,13 @@
 from pytest_mock import MockerFixture
 
 from .base_test import BaseTest
-from .mockers import (assert_failure, assert_success, mock_input_returning,
-                      mock_input_returning_y, read_branch_layout_file,
-                      rewrite_branch_layout_file)
-from .mockers_git_repository import (check_out, commit, create_repo,
-                                     create_repo_with_remote, delete_branch,
-                                     get_commit_hash, get_current_commit_hash,
-                                     new_branch, push)
+from .cli_runner import (assert_failure, assert_success,
+                         read_branch_layout_file, rewrite_branch_layout_file)
+from .git_repository import (check_out, commit, create_repo,
+                             create_repo_with_remote, delete_branch,
+                             get_commit_hash, get_current_commit_hash,
+                             new_branch, push)
+from .mockers import mock_input_returning, mock_input_returning_y
 
 
 class TestAdd(BaseTest):
