+++ tests/test_advance.py
@@ -7,15 +7,14 @@ from git_machete.utils import (ExitCode, FullTerminalAnsiOutputCodes,
                                UnderlyingGitException)
 
 from .base_test import BaseTest
-from .mockers import (assert_failure, assert_success, launch_command,
-                      launch_command_capturing_output_and_exception,
-                      mock_input_returning, mock_input_returning_y,
-                      rewrite_branch_layout_file,
-                      wait_to_bump_commit_timestamp)
-from .mockers_git_repository import (check_out, commit, create_repo,
-                                     create_repo_with_remote, fetch,
-                                     get_commit_hash, get_current_commit_hash,
-                                     new_branch, push, reset_to)
+from .cli_runner import (assert_failure, assert_success, launch_command,
+                         launch_command_capturing_output_and_exception,
+                         rewrite_branch_layout_file)
+from .git_repository import (check_out, commit, create_repo,
+                             create_repo_with_remote, fetch, get_commit_hash,
+                             get_current_commit_hash, new_branch, push,
+                             reset_to, wait_to_bump_commit_timestamp)
+from .mockers import mock_input_returning, mock_input_returning_y
 
 
 class TestAdvance(BaseTest):
