+++ tests/test_clean.py
@@ -1,11 +1,12 @@
 from pytest_mock import MockerFixture
 
 from .base_test import BaseTest
-from .mockers import (assert_success, launch_command, mock_input_returning_y,
-                      read_branch_layout_file, rewrite_branch_layout_file)
-from .mockers_git_repository import (add_remote, check_out, commit,
-                                     create_repo, create_repo_with_remote,
-                                     get_local_branches, new_branch, push)
+from .cli_runner import (assert_success, launch_command,
+                         read_branch_layout_file, rewrite_branch_layout_file)
+from .git_repository import (add_remote, check_out, commit, create_repo,
+                             create_repo_with_remote, get_local_branches,
+                             new_branch, push)
+from .mockers import mock_input_returning_y
 from .mockers_github import mock_github_token_for_domain_none
