+++ tests/test_fork_point.py
@@ -1,15 +1,15 @@
 import os
 
 from .base_test import BaseTest
-from .mockers import (assert_failure, assert_success, execute,
-                      fixed_author_and_committer_date_in_past, launch_command,
-                      remove_directory, rewrite_branch_layout_file)
-from .mockers_git_repository import (add_remote, check_out, commit,
-                                     create_repo, create_repo_with_remote,
-                                     delete_branch, fetch, get_commit_hash,
-                                     get_current_commit_hash, new_branch,
-                                     new_orphan_branch, pull, push, reset_to,
-                                     set_git_config_key)
+from .cli_runner import (assert_failure, assert_success, launch_command,
+                         rewrite_branch_layout_file)
+from .git_repository import (add_remote, check_out, commit, create_repo,
+                             create_repo_with_remote, delete_branch, fetch,
+                             get_commit_hash, get_current_commit_hash,
+                             new_branch, new_orphan_branch, pull, push,
+                             reset_to, set_git_config_key)
+from .mockers import fixed_author_and_committer_date_in_past
+from .shell import execute, remove_directory
 
 
 class TestForkPoint(BaseTest):
