+++ tests/test_github_sync.py
@@ -1,12 +1,12 @@
 from pytest_mock import MockerFixture
 
 from tests.base_test import BaseTest
-from tests.mockers import (assert_success, launch_command,
-                           mock_input_returning_y, rewrite_branch_layout_file)
+from tests.cli_runner import (assert_success, launch_command,
+                              rewrite_branch_layout_file)
+from tests.git_repository import (check_out, commit, create_repo_with_remote,
+                                  get_local_branches, new_branch, push)
+from tests.mockers import mock_input_returning_y
 from tests.mockers_code_hosting import mock_from_url
-from tests.mockers_git_repository import (check_out, commit,
-                                          create_repo_with_remote,
-                                          get_local_branches, new_branch, push)
 from tests.mockers_github import (MockGitHubAPIState,
                                   mock_github_token_for_domain_fake,
                                   mock_pr_json, mock_urlopen)
