+++ tests/test_github_update_pr_descriptions.py
@@ -4,13 +4,12 @@ from typing import Any, Dict, List
 from pytest_mock import MockerFixture
 
 from tests.base_test import BaseTest
-from tests.mockers import (assert_failure, assert_success, launch_command,
-                           rewrite_branch_layout_file)
+from tests.cli_runner import (assert_failure, assert_success, launch_command,
+                              rewrite_branch_layout_file)
+from tests.git_repository import (check_out, commit, create_repo_with_remote,
+                                  delete_branch, new_branch, push,
+                                  set_git_config_key)
 from tests.mockers_code_hosting import mock_from_url
-from tests.mockers_git_repository import (check_out, commit,
-                                          create_repo_with_remote,
-                                          delete_branch, new_branch, push,
-                                          set_git_config_key)
 from tests.mockers_github import (MockGitHubAPIState,
                                   mock_github_token_for_domain_fake,
                                   mock_github_token_for_domain_none,
