+++ tests/test_gitlab_create_mr.py
@@ -4,22 +4,22 @@ import textwrap
 from pytest_mock import MockerFixture
 
 from tests.base_test import BaseTest
-from tests.mockers import (assert_failure, assert_success, execute,
-                           fixed_author_and_committer_date_in_past,
-                           launch_command, mock_input_returning,
-                           mock_input_returning_y, rewrite_branch_layout_file,
-                           wait_to_bump_commit_timestamp, write_to_file)
+from tests.cli_runner import (assert_failure, assert_success, launch_command,
+                              rewrite_branch_layout_file)
+from tests.git_repository import (add_remote, amend_commit, check_out, commit,
+                                  create_repo, create_repo_with_remote,
+                                  delete_branch, delete_remote_branch,
+                                  new_branch, push, remove_remote, reset_to,
+                                  set_git_config_key,
+                                  wait_to_bump_commit_timestamp)
+from tests.mockers import (fixed_author_and_committer_date_in_past,
+                           mock_input_returning, mock_input_returning_y)
 from tests.mockers_code_hosting import mock_from_url
-from tests.mockers_git_repository import (add_remote, amend_commit, check_out,
-                                          commit, create_repo,
-                                          create_repo_with_remote,
-                                          delete_branch, delete_remote_branch,
-                                          new_branch, push, remove_remote,
-                                          reset_to, set_git_config_key)
 from tests.mockers_gitlab import (MockGitLabAPIState,
                                   mock_gitlab_token_for_domain_fake,
                                   mock_gitlab_token_for_domain_none,
                                   mock_mr_json, mock_urlopen)
+from tests.shell import execute, write_to_file
 
 
 class TestGitLabCreateMR(BaseTest):
