+++ tests/test_is_managed.py
@@ -1,8 +1,8 @@
 import pytest
 
 from .base_test import BaseTest
-from .mockers import launch_command, rewrite_branch_layout_file
-from .mockers_git_repository import check_out, commit, create_repo, new_branch
+from .cli_runner import launch_command, rewrite_branch_layout_file
+from .git_repository import check_out, commit, create_repo, new_branch
 
 
 class TestIsManaged(BaseTest):
