+++ tests/test_list.py
@@ -1,9 +1,8 @@
 from .base_test import BaseTest
-from .mockers import (assert_failure, assert_success, launch_command,
-                      rewrite_branch_layout_file)
-from .mockers_git_repository import (check_out, commit,
-                                     create_repo_with_remote, delete_branch,
-                                     new_branch, push)
+from .cli_runner import (assert_failure, assert_success, launch_command,
+                         rewrite_branch_layout_file)
+from .git_repository import (check_out, commit, create_repo_with_remote,
+                             delete_branch, new_branch, push)
 
 
 class TestList(BaseTest):
