+++ tests/test_log.py
@@ -1,10 +1,11 @@
 from pytest_mock import MockerFixture
 
 from .base_test import BaseTest
-from .mockers import (assert_success, fixed_author_and_committer_date_in_past,
-                      launch_command, mock__run_cmd_and_forward_stdout)
-from .mockers_git_repository import (commit, create_repo,
-                                     get_current_commit_hash, new_branch)
+from .cli_runner import assert_success, launch_command
+from .git_repository import (commit, create_repo, get_current_commit_hash,
+                             new_branch)
+from .mockers import (fixed_author_and_committer_date_in_past,
+                      mock__run_cmd_and_forward_stdout)
 
 
 class TestLog(BaseTest):
