+++ tests/test_show.py
@@ -1,11 +1,12 @@
 import os
 
 from .base_test import BaseTest
-from .mockers import (assert_failure, assert_success, execute, launch_command,
-                      remove_directory, rewrite_branch_layout_file)
-from .mockers_git_repository import (check_out, commit, create_repo,
-                                     create_repo_with_remote, new_branch,
-                                     new_orphan_branch, pull, push)
+from .cli_runner import (assert_failure, assert_success, launch_command,
+                         rewrite_branch_layout_file)
+from .git_repository import (check_out, commit, create_repo,
+                             create_repo_with_remote, new_branch,
+                             new_orphan_branch, pull, push)
+from .shell import execute, remove_directory
 
 
 class TestShow(BaseTest):
